Move DataFrame to submodule
